Created a custom hook for notifications
